Cleanup history
History before this point will now be a little wonky.

Windsock started off being developed in the shotover repo:
https://github.com/shotover/shotover-proxy
But as of this commit development is continuing in a seperate repo.

I ran `git-filter-repo --path windsock` to clean out all shotover
specific history.
And then I reintroduced a few files such as the workspace Cargo.toml in
this commit.

This comes at the cost of the previous commits being unbuildable.
But this project is not large enough to need reproducible commits
at this point in time.
And I think removing the 1000+ shotover specific commits is worth it.